+++ mysys/thr_mutex.c
@@ -512,7 +512,7 @@ int safe_cond_wait(pthread_cond_t *cond, safe_mutex_t *mp, const char *file,
     fprintf(stderr,
 	    "safe_mutex:  Count was %d in thread 0x%lx when locking mutex %s "
             "at %s, line %d\n",
-	    mp->count-1, my_thread_dbug_id(), mp->name, file, line);
+	    mp->count-1, (ulong) my_thread_dbug_id(), mp->name, file, line);
     fflush(stderr);
     abort();
   }
@@ -566,7 +566,7 @@ int safe_cond_timedwait(pthread_cond_t *cond, safe_mutex_t *mp,
     fprintf(stderr,
 	    "safe_mutex:  Count was %d in thread 0x%lx when locking mutex "
             "%s at %s, line %d (error: %d (%d))\n",
-	    mp->count-1, my_thread_dbug_id(), mp->name, file, line,
+	    mp->count-1, (ulong) my_thread_dbug_id(), mp->name, file, line,
             error, error);
     fflush(stderr);
     abort();
