+++ storage/spider/spd_db_conn.cc
@@ -694,7 +694,8 @@ int spider_db_errorno(
             "to %ld: %d %s\n",
             l_time->tm_year + 1900, l_time->tm_mon + 1, l_time->tm_mday,
             l_time->tm_hour, l_time->tm_min, l_time->tm_sec,
-            current_thd->thread_id, error_num, conn->db_conn->get_error());
+            (ulong) current_thd->thread_id, error_num,
+                  conn->db_conn->get_error());
         }
         if (!conn->mta_conn_mutex_unlock_later)
         {
@@ -714,7 +715,8 @@ int spider_db_errorno(
           "to %ld: %d %s\n",
           l_time->tm_year + 1900, l_time->tm_mon + 1, l_time->tm_mday,
           l_time->tm_hour, l_time->tm_min, l_time->tm_sec,
-          current_thd->thread_id, error_num, conn->db_conn->get_error());
+          (ulong) current_thd->thread_id, error_num,
+                conn->db_conn->get_error());
       }
       if (!conn->mta_conn_mutex_unlock_later)
       {
